move creep roles into separate folder
