+++ src/controller/creeps/role_base_transferor.js
@@ -1,9 +1,9 @@
-import { pickUpNearbyDroppedEnergy, moveAndWithdraw, moveAndTransfer, workingStatesKeeper, getEnergyFromNearbyLink, getEnergyFromStorage, setDoing } from '@/utils/util_beheavor'
+import { moveAndWithdraw, moveAndTransfer, getEnergyFromStorage, setDoing } from '@/utils/util_beheavor'
 
-import { task_powerSpawn } from './task_powerSpawn'
+import { task_powerSpawn } from '@/task_powerSpawn'
 
 import { RM } from '@/utils/util_consts'
-import task_factory from './task_factory'
+import task_factory from '@/task_factory'
