+++ src/controller/creeps/role_carrier.js
