+++ src/controller/creeps/role_guardian.js
