+++ src/controller/creeps/role_miner.js
