+++ src/controller/creeps/role_remote_carrier.js
