+++ src/controller/creeps/role_remote_claimer.js
