+++ src/controller/creeps/role_remote_harvester.js
