+++ src/controller/creeps/role_scavenger.js
