+++ src/controller/creeps/role_useless.js
